requirements.txt
pytest==7.1.3
selenium==4.5.0
webdriver-manager==3.8.3
Pillow==9.2.0
numpy==1.23.4
pytest-xdist==3.0.2
allure-pytest==2.11.1
flask==2.2.2
Werkzeug==2.2.2
SQLAlchemy==1.4.44
PyMySQL==1.0.2
cryptography==38.0.3
